Modify AuthContext
