+++ gptube/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import { openSans } from "../Common/Fonts";
-import { useAuth } from "@/context/AuthContext";
+import { useAuth } from "@/context/AuthContext/AuthContext";
 import { AiOutlineTwitter, AiOutlineGithub } from "react-icons/ai";
 import { useRouter } from "next/router";
 import Image from "next/image";
