+++ gptube/components/Security/ProtectedRoute.tsx
@@ -1,24 +1,23 @@
-import { useAuth, localStoreAuthVar } from "@/context/AuthContext";
-import { useRouter } from "next/router";
-import { ReactNode, useEffect } from "react";
-const ProtectedRoute: React.FC<{ children: ReactNode }> = ({ children }) => {
-  const { user } = useAuth();
-  const router = useRouter();
-  const { from } = router.query;
+import { useRouter } from 'next/router'
+import { useEffect } from 'react'
+
+import { useAuth } from '@/context/AuthContext/AuthContext'
+function ProtectedRoute({ children }: { children: React.ReactNode }) {
+  const { user } = useAuth()
+  const router = useRouter()
 
   useEffect(() => {
-    const isAuthenticated = localStorage.getItem(localStoreAuthVar);
-    if (!user && !isAuthenticated) {
+    if (!user) {
       router.push({
-        pathname: "/login",
+        pathname: '/login',
         query: {
-          from: from,
+          from: router.query.from,
         },
-      });
-    } else {
-      localStorage.setItem(localStoreAuthVar, "true");
+      })
     }
-  }, [router, user]);
-  return <>{user ? children : null}</>;
-};
-export default ProtectedRoute;
+  }, [router, user])
+
+  return <>{user ? children : null}</>
+}
+
+export default ProtectedRoute
