+++ gptube/pages/_app.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import "@/styles/globals.css";
 import { MyAppProps } from "@/components/Common/Types";
 import { Layouts } from "@/components/Layouts/Layouts";
-import { AuthContextProvider } from "@/context/AuthContext";
+import { AuthContextProvider } from "@/context/AuthContext/AuthContext";
 import { useRouter } from "next/router";
 import Script from "next/script";
 import ProtectedRoute from "@/components/Security/ProtectedRoute";
@@ -12,7 +12,7 @@ const noAuthRequired = ["/", "/login", "/reset", "/pricing"];
 export default function App({ Component, pageProps }: MyAppProps) {
   const router = useRouter();
   const Layout = Layouts[Component.Layout] ?? ((page) => page);
-  
+
   return (
     <>
       <AuthContextProvider>
