+++ gptube/pages/youtube/[resultID].tsx
@@ -13,7 +13,7 @@ import RobertaResultsBanner from '@/components/SocialMedia/Youtube/RobertaResult
 import { openSans } from '@/components/Common/Fonts'
 import NegativeComments from '@/components/SocialMedia/Youtube/NegativeComments'
 import { paramValToString } from '@/utils'
-import { useAuth } from '@/context/AuthContext'
+import { useAuth } from '@/context/AuthContext/AuthContext'
 import { firestore } from '@/lib/firebase/config-firebase'
 
 const { Content } = Layout
